+++ folded/LombokTestData-folded.java
@@ -431,4 +431,20 @@ import java.util.logging.Logger;
         }
     }
 
+    class Constructors {
+        @NoArgsConstructor @AllArgsConstructor public class FiveConstructors {
+            @Constructor(1) @Constructor(2) @Constructor(3) private int field1;
+            @Constructor(2) @Constructor(3) private String field2;
+            @Constructor(3) private double field3;
+            private boolean field4;
+        }
+
+        @NoArgsConstructor @AllArgsConstructor public class FiveConstructorsMod {
+            @Constructor(1,private) @Constructor(2,default) @Constructor(3,protected) private int field1;
+            @Constructor(2,default) @Constructor(3,protected) private String field2;
+            @Constructor(3,protected) private double field3;
+            private boolean field4;
+        }
+    }
+
 }